\name{searchapi-package}
\alias{searchapi-package}
\alias{searchapi}
\docType{package}
\title{
\packageTitle{searchapi}
}
\description{
\packageDescription{searchapi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{searchapi}
\packageIndices{searchapi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{searchapi}

Maintainer: \packageMaintainer{searchapi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
